split Bounty and Post rendering
With a file dedicated to Bounties, reward is not a concept in the layout
of a regular post. This is in anticipation of special handling of more
use cases such as WalletReviews which is also included here.
